+++ tests/test_main.py
@@ -3,6 +3,7 @@ import csv
 import json
 from tempfile import NamedTemporaryFile
 
+import pytest
 import responses
 from fastapi.testclient import TestClient
 
@@ -73,16 +74,16 @@ def test_csv_2_json_malformed_file():
     # GIVEN a POST request to the endpoint with multipart-encoded files:
     # (https://requests.readthedocs.io/en/latest/user/advanced/#post-multiple-multipart-encoded-files)
     files = [
-        ("files", (variants_hgvs_csv, open(variants_hgvs_csv_path, "rb"))),
+        ("files", (variants_hgvs_csv, open(casedata_snv_csv_path, "rb"))),  # files are switched
         (
             "files",
-            (casedata_old_csv, open(casedata_old_csv_path, "rb")),
+            (casedata_snv_csv, open(variants_hgvs_csv_path, "rb")),
         ),
     ]
     response = client.post("/csv_2_json", files=files)
     # THEN the endpoint should return error
     assert response.status_code == 400
-    assert "Created json file contains validation errors" in response.json()["message"]
+    assert response.json()["message"]
 
 
 def test_csv_2_json_old_format():
@@ -186,7 +187,10 @@ def test_csv_2_json_SV_breakpoints():
     # GIVEN a POST request to the endpoint with multipart-encoded files:
     # (https://requests.readthedocs.io/en/latest/user/advanced/#post-multiple-multipart-encoded-files)
     files = [
-        ("files", (variants_sv_breakpoints_csv, open(variants_sv_breakpoints_csv_path, "rb"))),
+        (
+            "files",
+            (variants_sv_breakpoints_csv, open(variants_sv_breakpoints_csv_path, "rb")),
+        ),
         ("files", (casedata_sv_csv, open(casedata_sv_csv_path, "rb"))),
     ]
 
@@ -216,7 +220,13 @@ def test_csv_2_json_SV_range_coords():
     # GIVEN a POST request to the endpoint with multipart-encoded files:
     # (https://requests.readthedocs.io/en/latest/user/advanced/#post-multiple-multipart-encoded-files)
     files = [
-        ("files", (variants_sv_range_coords_csv, open(variants_sv_range_coords_csv_path, "rb"))),
+        (
+            "files",
+            (
+                variants_sv_range_coords_csv,
+                open(variants_sv_range_coords_csv_path, "rb"),
+            ),
+        ),
         ("files", (casedata_sv_csv, open(casedata_sv_csv_path, "rb"))),
     ]
 
@@ -233,7 +243,14 @@ def test_csv_2_json_SV_range_coords():
     subm_coords = json_resp["clinvarSubmission"][0]["variantSet"]["variant"][0][
         "chromosomeCoordinates"
     ]
-    for item in ["assembly", "chromosome", "innerStart", "innerStop", "outerStart", "outerStop"]:
+    for item in [
+        "assembly",
+        "chromosome",
+        "innerStart",
+        "innerStop",
+        "outerStart",
+        "outerStop",
+    ]:
         assert item in subm_coords
     assert json_resp["clinvarSubmission"][0]["variantSet"]["variant"][0]["variantType"]
     assert json_resp["clinvarSubmission"][0]["variantSet"]["variant"][0]["referenceCopyNumber"]