+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [unreleased]
+### Changed
+- Updated submission schema to the latest API schema (01-03-2024)
+- Switched from Draft3Validator to Draft7Validator in validate.py
+
 ## [2.4]
 ### Changed
 - Updated version of external images in GitHub actions